% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post_insights.R
\name{cc_api_get_fb_page_post_insights}
\alias{cc_api_get_fb_page_post_insights}
\title{Get information about a single media directly from the API. Mostly used
internally.}
\usage{
cc_api_get_fb_page_post_insights(
  fb_post_id,
  metrics = cc_valid_fields_fb_post_insights,
  cache = TRUE,
  update = TRUE,
  api_version = "v18.0",
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\arguments{
\item{metrics}{Metrics to be retrieved. Consider that depending on the media
type, different media types are effectively available. Requesting the wrong
metrics will cause an error. Defaults to NULL. If left to NULL, metrics will be chosen based on the media type. See the official documentation for reference:
\url{https://developers.facebook.com/docs/instagram-api/reference/ig-media/insights}}

\item{api_version}{Defaults to "v18.0".}

\item{ig_media_id}{Instagram media identifier, must be a vector of length 1.
A list of identifiers for your account can be retrieved with
`cc_get_instagram_media_id()`.}

\item{media_type}{Media type. Valid values include "IMAGE", "VIDEO", "REELS", and
"CAROUSEL_ALBUM". Defaults to NULL. If not given, it will be retrieved with
`cc_get_instagram_media`. Ignored if `metrics` explicitly given.}
}
\description{
See the official documentation for reference:
\url{https://developers.facebook.com/docs/instagram-api/reference/ig-media/insights}
}
