% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_check_ig_media_update.R
\name{cc_check_instagram_media_update}
\alias{cc_check_instagram_media_update}
\title{Check when information about Instagram media should be updated based on a simple heuristic}
\usage{
cc_check_instagram_media_update(
  ig_media_id = NULL,
  ig_user_id = NULL,
  insights = FALSE,
  token = NULL
)
}
\arguments{
\item{ig_media_id}{Instagram media identifier. A list of identifiers for your
account can be retrieved with `cc_get_instagram_media_id()`. If left to
NULL, a full list is automatically retrieved.}

\item{ig_user_id}{Instagram user id, typically composed of 17 digits. Not to
be confused with legacy Instragram account id.}
}
\value{
A tibble with four columns: `ig_media_id` (a character column),
  `date_created` and `date_retrieved` (Date columns), and an `update` column
  (logical)
}
\description{
Update is set to TRUE if:
}
\details{
- media has been published in the last week and not retrieved today
- media has been published in the last month and not retrieved in the last week
- media has been published in the last year and not retrieved in the last month
- media has not been retrieved in the last year
}
