% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_date_to_json.R
\name{cc_date_to_json}
\alias{cc_date_to_json}
\title{Output date in a format that can be passed as custom date to the Facebok API.}
\usage{
cc_date_to_json(start_date, end_date = NULL)
}
\arguments{
\item{start_date}{A date.}

\item{end_date}{Defaults to NULL. If left to NULL, `end_date` is assumed to
be the same as the start date}
}
\description{
Output date in a format that can be passed as custom date to the Facebok API.
}
