% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_download_facebook_media.R
\name{cc_download_facebook_media}
\alias{cc_download_facebook_media}
\title{Downloads media or previews of Facebook posts}
\usage{
cc_download_facebook_media(
  posts_df = NULL,
  type = c("full_picture"),
  path = NULL,
  wait = 1,
  ...
)
}
\arguments{
\item{posts_df}{Defaults to `NULL`. If given, expected to be a data frame
generated with [cc_get_fb_page_posts()].}

\item{type}{Defaults to "full_picture", type of media to download. Other
options not yet implemented.}

\item{path}{Path to the folder where files will be downloaded; defaults to
`fb_full_picture`, depending on `type`. If the path does
not exist, it will be created.}

\item{wait}{Defaults to 1. Time to wait between downloading each file.}

\item{...}{Passed to [cc_get_fb_page_posts()].}
}
\value{
Nothing, used for its side effects.
}
\description{
Downloads media or previews of Facebook posts
}
\examples{
\dontrun{
if (interactive()) {
   cc_download_instagram_media()
}
}
}
