% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_download_instagram_media.R
\name{cc_download_instagram_media}
\alias{cc_download_instagram_media}
\title{Downloads Instagram media or thumbnails}
\usage{
cc_download_instagram_media(
  media_df = NULL,
  type = c("media", "thumbnail"),
  path = NULL,
  wait = 1,
  ...
)
}
\arguments{
\item{media_df}{Defaults to `NULL`. If given, expected to be a data frame generated with [cc_get_instagram_media()].}

\item{type}{Defaults to "media", type of media to download. The only alternative valid value is "thumbnail".}

\item{path}{Path to the folder where files will be downloaded; defaults to either `ig_media` or `ig_thumbnail`, depending on `type`. If the path does not exist, it will be created.}

\item{wait}{Defaults to 1. Time to wait between downloading each file.}

\item{...}{Passed to [cc_get_instagram_media()].}
}
\value{
Nothing, used for its side effects.
}
\description{
Downloads Instagram media or thumbnails
}
\examples{
\dontrun{
if (interactive()) {
   cc_download_instagram_media()
}
}
}
