% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_linkedin_stats_content.R
\name{cc_drive_upload_linkedin_stats_content}
\alias{cc_drive_upload_linkedin_stats_content}
\title{Upload to Google Sheets Metrics statistics exported from LinkedIn}
\usage{
cc_drive_upload_linkedin_stats_content(
  path,
  page_name,
  sheet_dribble = NULL,
  base_dribble = NULL,
  export_csv = FALSE
)
}
\arguments{
\item{path}{Base path where all xls files exported from LinkedIn are stored.}

\item{page_name}{Name of the page. If unsure about the exact form, see the
relevant column of `cc_get_linkedin_stats_files()`}

\item{sheet_dribble}{Dribble of the specific sheet where files should be uploaded.}

\item{base_dribble}{Dribble of the folder where the relevant sheet is expected to be.}

\item{export_csv}{Defaults to TRUE. If TRUE, exports content stats in a csv
file in a folder with the same name as the base path, but with "_processed"
appended.}
}
\description{
Upload to Google Sheets Metrics statistics exported from LinkedIn
}
