% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ad_actions_by_day.R
\name{cc_get_fb_ad_actions_by_day}
\alias{cc_get_fb_ad_actions_by_day}
\title{Takes all action results for all days when an ad is active}
\usage{
cc_get_fb_ad_actions_by_day(
  ad_id = NULL,
  type = "actions",
  cache = TRUE,
  only_cached = FALSE,
  api_version = "v17.0",
  ad_account_id = NULL,
  token = NULL
)
}
\description{
Draft: no caching and returns only 25 days for each ad, as paging in data retrieval has not been enabled
}
