% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ad_creatives_id.R
\name{cc_get_fb_ad_creatives_id}
\alias{cc_get_fb_ad_creatives_id}
\title{Get Facebook ad creatives id for each ad id}
\usage{
cc_get_fb_ad_creatives_id(
  ad_id = NULL,
  ad_account_id = NULL,
  api_version = "v19.0",
  cache = TRUE,
  token = NULL
)
}
\value{
A data frame with three character columns: `ad_id`, `creative_id`, and `timestamp_retrieved`
}
\description{
See: https://developers.facebook.com/docs/marketing-api/reference/adgroup/adcreatives/
}
