% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ad_custom.R
\name{cc_get_fb_ad_custom}
\alias{cc_get_fb_ad_custom}
\title{Get custom stats about an ad (currently not functional)}
\usage{
cc_get_fb_ad_custom(
  fields = "reach",
  breakdowns = "frequency_value",
  ad_id = NULL,
  start_date = NULL,
  end_date = NULL,
  only_cached = FALSE,
  meta_api_version = cornucopia::cc_get_meta_api_version(),
  cache = TRUE,
  ad_account_id = NULL,
  fb_user_token = NULL
)
}
\description{
https://developers.facebook.com/docs/marketing-api/reference/ad-campaign-stats/
}
\details{
attribution window
https://developers.facebook.com/docs/marketing-api/reference/ads-action-stats/
}
