% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ad_frequency_distribution.R
\name{cc_get_fb_ad_frequency_distribution}
\alias{cc_get_fb_ad_frequency_distribution}
\title{Get frequency distribution}
\usage{
cc_get_fb_ad_frequency_distribution(
  ad_id = NULL,
  date_preset = c("maximum", "last_7d", "last_30d", "today", "yesterday", "this_month",
    "last_month", "this_quarter", "maximum", "data_maximum", "last_3d", "last_7d",
    "last_14d", "last_28d", "last_30d", "last_90d", "last_week_mon_sun",
    "last_week_sun_sat", "last_quarter", "last_year", "this_week_mon_today",
    "this_week_sun_today", "this_year"),
  output = c("tibble", "barchart", "list"),
  start_date = NULL,
  end_date = NULL,
  meta_api_version = cornucopia::cc_get_meta_api_version(),
  ad_account_id = NULL,
  fb_user_token = NULL
)
}
\arguments{
\item{ad_id}{Indentifier of a an `ad_id`, `adset_id`, or `campaign_id`.}

\item{date_preset}{Defaults to "maximum". Available values include "last_7d",
"last_30d", and similar, as described in the
\href{https://developers.facebook.com/docs/marketing-api/insights/breakdowns}{official
API documentation}. N.B. This is ignored if both `start_date` and
`end_date` are given.}

\item{output}{Output format, defaults to `tibble`. Valid values include
`list` and `barchart`, which outputs a `ggplot2` object.}

\item{start_date}{The beginning date for the period to be considered. Both
`start_date` and `end_date` must be given. If either is ignored, `preset`
takes precedence.}

\item{end_date}{The end date for the period to be considered. Both
`start_date` and `end_date` must be given. If either is ignored, `preset`
takes precedence.}
}
\description{
For details, see the
\href{https://developers.facebook.com/docs/marketing-api/insights/breakdowns}{official
documentation on Insights breakdowns}.
}
