% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ads.R
\name{cc_get_fb_ads}
\alias{cc_get_fb_ads}
\title{Get details about Facebook ads}
\usage{
cc_get_fb_ads(start_date = NULL, end_date = NULL, only_cached = FALSE)
}
\arguments{
\item{only_cached}{Defaults to FALSE. If TRUE, only pre-cached files within
the given date range are loaded; no new calls to the API are made and
reliably works offline.}
}
\value{
A data frame with an extensive number of fields, some presented as
  nested data frames.
}
\description{
Caches data in the folder `fb_ads_by_date_rds` in the current working
directory.
}
\details{
See also `cc_get_fb_ads_by_date()` for customisation of fields.
}
\examples{
\dontrun{
cc_get_fb_ads()
}
}
