% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ads_campaigns.R
\name{cc_get_fb_ads_campaigns}
\alias{cc_get_fb_ads_campaigns}
\title{Get all campaigns for the current ad account}
\usage{
cc_get_fb_ads_campaigns(
  api_version = "v17.0",
  ad_account_id = NULL,
  token = NULL
)
}
\description{
For details on valid fields, see: https://developers.facebook.com/docs/marketing-api/reference/ad-campaign-group/
}
\details{
Draft: not yet caching or paging results
}
\examples{
\dontrun{
cc_get_fb_ads_campaigns()
}
}
