% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ads_daily.R
\name{cc_get_fb_ads_daily}
\alias{cc_get_fb_ads_daily}
\title{Get all campaigns for the current ad account / legacy}
\usage{
cc_get_fb_ads_daily(
  fields = c("spend"),
  level = "campaign",
  start_date = NULL,
  end_date = NULL,
  api_version = "v19.0",
  fb_ad_account_id = NULL,
  max_pages = NULL,
  fb_user_token = NULL,
  cache = TRUE,
  update = TRUE
)
}
\arguments{
\item{fields}{A character vector of fields to retrieve. Defaults to "spend".}
}
\description{
For details on fields, see \url{https://developers.facebook.com/docs/marketing-api/reference/ad-campaign-group/}
}
\details{
It currently returns all fields that return a single value,
`cc_valid_fields_ad_campaign_group_v`

For reference, these are the fields that are always present for all
campaigns:
"id,name,created_time,updated_time,start_time,stop_time,objective,status"

Cache updating currently suboptimal.
}
\examples{
\dontrun{
cc_get_fb_ads_daily()
}
}
