% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_managed_pages.R
\name{cc_get_fb_managed_pages}
\alias{cc_get_fb_managed_pages}
\title{Get managed pages, including name, page token, and id}
\usage{
cc_get_fb_managed_pages(
  fields = c("id", "name"),
  fb_user_id = NULL,
  fb_user_token = NULL
)
}
\arguments{
\item{fields}{Defaults to `c("id", "name")`. Include "access_token" in order to retrieve you Facebook page access token.}

\item{fb_user_id}{Facebook user id. Defaults to NULL, can be set with
`cc_set()`. Can be retrieved with `cc_get_fb_user()`.}

\item{fb_user_token}{Facebook user token (not a page token).}
}
\description{
Get managed pages, including name, page token, and id
}
\examples{
\dontrun{
cc_get_fb_managed_pages()
}
}
