% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_insights.R
\name{cc_get_fb_page_insights}
\alias{cc_get_fb_page_insights}
\title{Title}
\usage{
cc_get_fb_page_insights(
  metric = c("page_impressions", "page_engaged_users"),
  api_version = "v18.0",
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\arguments{
\item{fb_page_token}{}
}
\description{
Official documentation: \url{https://developers.facebook.com/docs/graph-api/reference/v18.0/insights}
}
