% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post.R
\name{cc_get_fb_page_post}
\alias{cc_get_fb_page_post}
\title{Get Facebook page post id and basic information that can be stored as strings}
\usage{
cc_get_fb_page_post(
  api_version = "v18.0",
  max_pages = NULL,
  cache = TRUE,
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\arguments{
\item{fb_page_token}{}
}
\description{
Get Facebook page post id and basic information that can be stored as strings
}
