% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post_url.R
\name{cc_get_fb_page_post_comment_url}
\alias{cc_get_fb_page_post_comment_url}
\title{Extract url posted in comments}
\usage{
cc_get_fb_page_post_comment_url(
  comments_df,
  posts_df = NULL,
  only_own_comments = TRUE,
  fb_page_id = NULL
)
}
\arguments{
\item{comments_df}{A data frame created with [cc_get_fb_page_post_comments()].}

\item{posts_df}{A data frame created with [cc_get_fb_page_posts()].}
}
\value{
A data frame, similar to `comments_df`, but with an additional `url_message` column.
}
\description{
Extract url posted in comments
}
