% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post_comments.R
\name{cc_get_fb_page_post_comments}
\alias{cc_get_fb_page_post_comments}
\title{Get Facebook page posts comments}
\usage{
cc_get_fb_page_post_comments(
  fb_post_id = NULL,
  cache = TRUE,
  update = TRUE,
  meta_api_version = cornucopia::cc_get_meta_api_version(),
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\arguments{
\item{fb_post_id}{Facebook post identifier, must be a vector of length 1.
A list of identifiers for a given Facebook page can be retrieved with
[cc_get_fb_page_posts()].}

\item{cache}{Defaults to `TRUE`.}
}
\value{
A data frame with comments.
}
\description{
For reference, see the
\href{https://developers.facebook.com/docs/graph-api/reference/page-post/comments}{official
documentation}.
}
