% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post_url.R
\name{cc_get_fb_page_post_url}
\alias{cc_get_fb_page_post_url}
\title{Extract urls included in posts shared on Facebook pages}
\usage{
cc_get_fb_page_post_url(posts_df = NULL, ...)
}
\arguments{
\item{posts_df}{A data frame created with [cc_get_fb_page_posts()].}
}
\value{
A data frame, including an `url_share` column (with urls explicitly shared), `url_message` column for url shared in the main text of the post, and `url_all`, combining the previous two and dropping NA.
}
\description{
Extracts both urls explicitly shared, as well as those included in the message of the post.
}
\examples{
\dontrun{
url_all_df <- cc_get_fb_page_posts(only_cached = TRUE, process_json = TRUE) |>
  cc_get_fb_page_post_url()

url_all_df |>
  tidyr::unnest("url_all") |>
  dplyr::select(c("created_time", "id", "url_all"))
  }
}
