% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post_url.R
\name{cc_get_fb_page_post_url_combo}
\alias{cc_get_fb_page_post_url_combo}
\title{Retrieve urls posted in a Facebook page, no matter if as attachment, in the
message of a post, or in a comment}
\usage{
cc_get_fb_page_post_url_combo(posts_with_url_df, comments_with_url_df)
}
\arguments{
\item{posts_with_url_df}{A data frame created with
[cc_get_fb_page_post_url()].}

\item{comments_with_url_df}{A data frame created with
[cc_get_fb_page_post_comment_url()].}
}
\value{
A data frame, with url and reference to Facebook post identifier.
}
\description{
Retrieve urls posted in a Facebook page, no matter if as attachment, in the
message of a post, or in a comment
}
