% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post.R
\name{cc_get_fb_page_posts}
\alias{cc_get_fb_page_posts}
\title{Get Facebook page post id and basic information that can be stored as strings}
\usage{
cc_get_fb_page_posts(
  api_version = "v18.0",
  max_pages = NULL,
  cache = TRUE,
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\arguments{
\item{max_pages}{Posts are returned in pages of 25 posts each. How many pages should be retrieved? By default, this will try to retrieve all posts.}

\item{cache}{Defaults to TRUE.}
}
\description{
Get Facebook page post id and basic information that can be stored as strings
}
\examples{
\dontrun{
if (interactive) {
  fb_page_token <- cc_get_fb_page_token(
    fb_user_id = cc_get_fb_user(),
    page_name = "My example page"
  )

  cc_set(fb_page_token = fb_page_token)
  posts_df <- cc_get_fb_page_posts()
  posts_df
}
}
}
