% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_video.R
\name{cc_get_fb_page_video}
\alias{cc_get_fb_page_video}
\title{Get Facebook page video id and basic information that can be stored as strings}
\usage{
cc_get_fb_page_video(
  meta_api_version = cornucopia::cc_get_meta_api_version(),
  max_pages = NULL,
  cache = TRUE,
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\description{
See: https://developers.facebook.com/docs/video-api/guides/get-videos/
}
