% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_products.R
\name{cc_get_fb_products}
\alias{cc_get_fb_products}
\title{Get information about a Facebook catalog products}
\usage{
cc_get_fb_products(
  fb_product_catalog_id = NULL,
  fields = cc_valid_fields_fb_product,
  api_version = "v17.0",
  max_pages = NULL,
  token = NULL
)
}
\value{
A data frame. Some columns include nested data.
}
\description{
Data are not cached locally.
}
\examples{
\dontrun{
  cc_get_fb_products()
}
}
