% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_user.R
\name{cc_get_fb_user}
\alias{cc_get_fb_user}
\title{Get Facebook user id and name, as well as other options fields}
\format{
Defaults to "data.frame". If list, a list is returned instead; mostly
  useful when the "picture" field is requested.
}
\usage{
cc_get_fb_user(token = NULL, fields = c("id", "name"), format = "data.frame")
}
\arguments{
\item{fields}{Defaults to all available, consider reducing if you don't have all relevant permissions.}
}
\value{
By default, a data frame with one row and two character columns,
  "name" and "id".
}
\description{
For details, see:
\url{https://developers.facebook.com/docs/graph-api/reference/user/}
}
\examples{
\dontrun{
cc_get_fb_user()
}
}
