% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_video_insights.R
\name{cc_get_fb_video_insights}
\alias{cc_get_fb_video_insights}
\title{Get Facebook video insights}
\usage{
cc_get_fb_video_insights(
  fb_video_id,
  metrics = cornucopia::cc_valid_fields_fb_video_insights,
  cache = TRUE,
  update = TRUE,
  meta_api_version = cornucopia::cc_get_meta_api_version(),
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\description{
For reference, see: https://developers.facebook.com/docs/graph-api/reference/video/video_insights/
defaults to Lifetime period for each video
}
