% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_bd_user.R
\name{cc_get_instagram_bd_user}
\alias{cc_get_instagram_bd_user}
\title{Retrieve information about a user through Instagram's business discovery}
\usage{
cc_get_instagram_bd_user(
  ig_username,
  fields = c("id", "ig_id", "username", "name", "biography", "website",
    "followers_count", "follows_count", "media_count", "profile_picture_url"),
  api_version = "v21.0",
  ig_user_id = NULL,
  fb_user_token = NULL
)
}
\arguments{
\item{ig_username}{A user name of an Instagram user.}

\item{fields}{Defaults to all fields publicly available through `business_discovery`. See [the documentation](https://developers.facebook.com/docs/instagram-platform/reference/instagram-media) for other fields that may be available.}
}
\description{
This function allows you to retrieve basic information about an Instagram account you are not associated with, as long as they are a business account.
}
\details{
[More details about Business Discovery and relevant permissions](https://developers.facebook.com/docs/instagram-platform/instagram-graph-api/reference/ig-user/business_discovery/).
}
\examples{
\dontrun{
if (interactive) {
  # e.g. to retrieve information about the Instagram account of the United Nations
  cc_get_instagram_bd_user(ig_username = "unitednations")
}
}
}
