% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_bd_user.R
\name{cc_get_instagram_bd_user_basic}
\alias{cc_get_instagram_bd_user_basic}
\title{Retrieve identifier of Instagram user}
\usage{
cc_get_instagram_bd_user_basic(
  ig_username,
  cache = TRUE,
  meta_api_version = cornucopia::cc_get_meta_api_version(),
  ig_user_id = NULL,
  fb_user_token = NULL
)
}
\arguments{
\item{ig_username}{Instagram username of an Instagram business user.}
}
\description{
Mostly used internally for consistent storing and caching.
}
\examples{
\dontrun{
if (interactive) {
  # e.g. to retrieve information about the Instagram account of the United Nations
  cc_get_instagram_bd_user_basic(ig_username = "unitednations")
}
}
}
