% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_bd_user_media.R
\name{cc_get_instagram_bd_user_media}
\alias{cc_get_instagram_bd_user_media}
\title{Retrieve information about other users through `business_discovery`}
\usage{
cc_get_instagram_bd_user_media(
  ig_username,
  fields = c("username", "timestamp", "like_count", "comments_count", "caption",
    "media_product_type", "media_type", "media_url", "thumbnail_url", "permalink"),
  max_pages = NULL,
  update = TRUE,
  cache = TRUE,
  api_version = "v21.0",
  ig_user_id = NULL,
  fb_user_token = NULL
)
}
\arguments{
\item{ig_username}{A user name of an Instagram user.}

\item{fields}{Defaults to all fields publicly available through `business_discovery`. See [the documentation](https://developers.facebook.com/docs/instagram-platform/reference/instagram-media) for other fields that may be available.}

\item{max_pages}{Posts are returned in pages of 25 posts each. How many pages should be retrieved? By default, this will try to retrieve all posts.}

\item{cache}{Defaults to TRUE.}
}
\description{
Consider that only information about posts of creative or business users may be available.
Given restrictions on the rate limit, you are likely to hit rate limits quite soon.
Wait one hour and try again.
}
\details{
For details about rate limits, see [this section of the documentation](https://developers.facebook.com/docs/graph-api/overview/rate-limiting).

[More details about Business Discovery and relevant permissions](https://developers.facebook.com/docs/instagram-platform/instagram-graph-api/reference/ig-user/business_discovery/).

In brief, necessary:

- `instagram_basic`
- `instagram_manage_insights`
- `pages_read_engagement` or `pages_show_list`

If the token is from a User whose Page role was granted via the Business Manager, one of the following permissions is also required:

- `ads_management`
- `pages_read_engagement`
- `business_management`
}
