% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_user.R
\name{cc_get_instagram_user}
\alias{cc_get_instagram_user}
\title{Get information about your Instagram user}
\usage{
cc_get_instagram_user(
  ig_user_id = NULL,
  meta_api_version = cornucopia::cc_get_meta_api_version(),
  fields = c("id", "ig_id", "username", "name", "biography", "website",
    "followers_count", "follows_count", "media_count", "profile_picture_url"),
  fb_user_token = NULL
)
}
\arguments{
\item{ig_user_id}{Instagram user id, typically composed of 17 digits. Not to
be confused with legacy Instagram account id.}

\item{meta_api_version}{Defaults to the latest API at the time this package was
last updated. Currently, this corresponds to api version 24.0.}

\item{fields}{Defaults to all available (except
"shopping_product_tag_eligibility", which requires dedicated permissions).
Consider reducing if you don't have all relevant permissions.}

\item{fb_user_token}{Facebook user token different from page token. Can be
retrieved with [cc_get_fb_user_token()] or [cc_get_fb_long_user_token()].}
}
\value{
A data frame (a tibble), with a column for id, plus as many columns
  as the requested fields.
}
\description{
If you need to retrieve your Instagram user id, consider
`cc_get_instagram_user_id()`.
}
\details{
See the relevant page in the documentation for available fields and more
details https://developers.facebook.com/docs/instagram-api/reference/ig-user

Look in particular at the permissions requirements. If you have issues,
consider dropping `shopping_product_tag_eligibility` from the fields, as it
requires additional permissions.
}
\examples{
\dontrun{
cc_get_instagram_user()
}
}
