% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_user_id.R
\name{cc_get_instagram_user_id}
\alias{cc_get_instagram_user_id}
\title{Get the Instagram user id of a Page's Instagram Business Account}
\usage{
cc_get_instagram_user_id(
  fb_page_id = NULL,
  fb_user_token = NULL,
  api_version = "v22.0"
)
}
\arguments{
\item{fb_user_token}{Facebook user token (not a page token).}
}
\description{
For details, see step 5 of this guide: \url{https://developers.facebook.com/docs/instagram-api/getting-started/}
}
