% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set.R
\name{cc_get_meta_api_version}
\alias{cc_get_meta_api_version}
\title{Retrieve version of Meta API.}
\usage{
cc_get_meta_api_version(meta_api_version = NULL)
}
\arguments{
\item{meta_api_version}{Defaults to the latest API at the time this package was
last updated. Currently, this corresponds to api version 24.0.}
}
\value{
The META API version set for the current session as a character
  vector of length 1.
}
\description{
Retrieve version of Meta API.
}
\examples{
cc_get_meta_api_version()

cc_get_meta_api_version("23.0")

cc_set(meta_api_version = "24.0")

cc_get_meta_api_version()

cc_get_meta_api_version("23.0")
}
