% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set.R
\name{cc_get_woocommerce_api_version}
\alias{cc_get_woocommerce_api_version}
\title{Retrieve version of Woocommerce API.}
\usage{
cc_get_woocommerce_api_version(woocommerce_api_version = NULL)
}
\description{
Retrieve version of Woocommerce API.
}
\examples{
cc_get_woocommerce_api_version()

cc_get_woocommerce_api_version("v3")

cc_set(woocommerce_api_version = "v2")

cc_get_woocommerce_api_version()

cc_set(woocommerce_api_version = "v3")
}
