% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set.R
\name{cc_get_woocommerce_base_url}
\alias{cc_get_woocommerce_base_url}
\title{Retrieve base url of Woocommerce API.}
\usage{
cc_get_woocommerce_base_url(woocommerce_base_url = NULL)
}
\value{
The base url as a character vector of length 1.
}
\description{
Typically it follows a pattern such as: `https://www.example.com/wp-json/wc/`.
Do not include the API version (e.g. `v3`), which is set separately.
}
\examples{
cc_get_woocommerce_base_url()

cc_get_woocommerce_base_url("https://www.example.com/wp-json/wc/")

cc_set(woocommerce_base_url = "https://www.example.com/custom/wp-json/wc/")

cc_get_woocommerce_base_url()
}
