% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_nomnoml.R
\name{cc_nomnoml}
\alias{cc_nomnoml}
\title{Show ad campaign structure in a nomnmonl graph}
\usage{
cc_nomnoml(ads_df, direction = "right", drop_vertical_bar = FALSE, ...)
}
\arguments{
\item{direction}{Defaults to "right". Can be "down". Passed to `nomnoml`.}

\item{drop_vertical_bar}{Defaults to FALSE. If TRUE, replaces vertical bar
(which in `nomnoml` can be used to separate a block in subsections) with
"-".}

\item{...}{}
}
\description{
Show ad campaign structure in a nomnmonl graph
}
