% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_read_fb_ads_by_date.R
\name{cc_read_fb_ads_by_date}
\alias{cc_read_fb_ads_by_date}
\title{Reads locally stored dated files, typically generated with `cc_get_fb_ads_by_date`}
\usage{
cc_read_fb_ads_by_date(
  path = "fb_ads_by_date_rds",
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{path}{Path to dated files}

\item{start_date}{Defaults to NULL. If given, only files retrieved on this date or later are included. Input should be of date class, or on the YYYY-MM-DD format.}

\item{end_date}{Defaults to NULL. If given, only files retrieved on this date or sooner are included. Input should be of date class, or on the YYYY-MM-DD format.}
}
\description{
Reads locally stored dated files, typically generated with `cc_get_fb_ads_by_date`
}
\examples{
\dontrun{
if (interactive()) {
  cc_read_fb_ads_by_date()
}
}
}
