% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_ui_instagram_user_card.R
\name{cc_ui_instagram_user_card}
\alias{cc_ui_instagram_user_card}
\title{Create a card to be used in dashboard or shiny app with basic information
about the Instagram account}
\usage{
cc_ui_instagram_user_card(
  ig_user_df = NULL,
  ig_user_id = NULL,
  fb_user_token = NULL
)
}
\arguments{
\item{ig_user_id}{Instagram user id, typically composed of 17 digits. Not to
be confused with legacy Instagram account id.}

\item{fb_user_token}{Facebook user token different from page token. Can be
retrieved with [cc_get_fb_user_token()] or [cc_get_fb_long_user_token()].}
}
\value{
A `bslib` card, to be used in Shiny apps.
}
\description{
Create a card to be used in dashboard or shiny app with basic information
about the Instagram account
}
\examples{
\dontrun{
if (interactive) {
  cc_ui_instagram_user_card()
}
}
}
