% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ad_actions_by_day.R
\name{cc_get_fb_ad_actions_by_day}
\alias{cc_get_fb_ad_actions_by_day}
\title{Takes all action results for all days when an ad is active}
\usage{
cc_get_fb_ad_actions_by_day(
  ad_id = NULL,
  type = "actions",
  action_breakdowns = NULL,
  cache = FALSE,
  only_cached = FALSE,
  api_version = "v22.0",
  ad_account_id = NULL,
  fb_user_token = NULL,
  max_pages = NULL
)
}
\arguments{
\item{only_cached}{Defaults to FALSE. If TRUE, only data cached locally will
be retrieved.}
}
\description{
For action breakdowns, see: https://developers.facebook.com/docs/marketing-api/insights/breakdowns/
}
\details{
Draft: caching disabled by default as only partly functional;
Not yet fully tested with ads running longer than 25 days
}
