% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ad_actions_by_day.R
\name{cc_get_fb_ad_actions_by_day}
\alias{cc_get_fb_ad_actions_by_day}
\title{Takes all action results for all days when an ad is active}
\usage{
cc_get_fb_ad_actions_by_day(
  ad_id = NULL,
  type = "actions",
  cache = FALSE,
  only_cached = FALSE,
  api_version = "v17.0",
  ad_account_id = NULL,
  token = NULL,
  max_pages = NULL
)
}
\description{
Draft: caching disabled by default as only partly functional;
Not yet fully tested with ads running longer than 25 days
}
