% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ad_custom.R
\name{cc_get_fb_ad_custom}
\alias{cc_get_fb_ad_custom}
\title{Get custom stats about an ad (currently not functional)}
\usage{
cc_get_fb_ad_custom(
  start_date = NULL,
  end_date = NULL,
  only_cached = FALSE,
  api_version = "v18.0",
  cache = TRUE,
  ad_account_id = NULL,
  token = NULL
)
}
\arguments{
\item{token}{}
}
\description{
https://developers.facebook.com/docs/marketing-api/reference/ad-campaign-stats/
}
\details{
attribution window
https://developers.facebook.com/docs/marketing-api/reference/ads-action-stats/
}
