% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_ads_campaigns.R
\name{cc_get_fb_ads_campaigns}
\alias{cc_get_fb_ads_campaigns}
\title{Get all campaigns for the current ad account}
\usage{
cc_get_fb_ads_campaigns(
  api_version = "v17.0",
  ad_account_id = NULL,
  max_pages = NULL,
  token = NULL,
  cache = FALSE
)
}
\description{
For details on fields, see: https://developers.facebook.com/docs/marketing-api/reference/ad-campaign-group/
}
\details{
It currently returns only fields that are always present for all campaigns: "id,name,created_time,updated_time,start_time,stop_time,objective,status"

Draft: paging results, but not yet caching
}
\examples{
\dontrun{
cc_get_fb_ads_campaigns()
}
}
