% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_long_page_token.R
\name{cc_get_fb_long_page_token}
\alias{cc_get_fb_long_page_token}
\title{Get a long-lived page acess token for Facebook}
\usage{
cc_get_fb_long_page_token(
  fb_user_id = NULL,
  fb_user_token = NULL,
  api_version = "v19.0"
)
}
\arguments{
\item{fb_user_id}{App-scoped user id. This can be retrieved with
`cc_get_fb_user()`}

\item{fb_user_token}{Must be a long-lived user token. This can be retrieved
with `cc_get_fb_long_user_token()`.}
}
\value{
A data frame with long-lived access tokens to all available pages.
}
\description{
For details, see
\url{https://developers.facebook.com/docs/facebook-login/guides/access-tokens/get-long-lived/}
}
\details{
You should use as input a long-lived user access token. Long-lived Page
access token do not have an expiration date and only expire or are
invalidated under certain conditions.
}
\examples{
\dontrun{
if (interactive) {
  cc_get_fb_long_page_token(
    fb_user_id = "your_fb_user_id_here",
    fb_user_token = "your_long_term_token_here"
  )
}
}

}
