% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_managed_pages.R
\name{cc_get_fb_managed_pages}
\alias{cc_get_fb_managed_pages}
\title{Get managed pages, including name, page token, and id}
\usage{
cc_get_fb_managed_pages(
  fb_user_id = NULL,
  fields = c("name", "access_token"),
  fb_user_token = NULL
)
}
\arguments{
\item{fb_user_id}{Facebook user id. Defaults to NULL, can be set with
`cc_set()`. Can be retrieved with `cc_get_fb_user()`.}

\item{fields}{Defaults to all available, consider reducing if you don't have all relevant permissions.}
}
\description{
Get managed pages, including name, page token, and id
}
\examples{
\dontrun{
cc_get_fb_managed_pages()
}
}
