% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_owned_product_catalogs.R
\name{cc_get_fb_owned_product_catalogs}
\alias{cc_get_fb_owned_product_catalogs}
\title{Get information about your Facebook product catalogs}
\usage{
cc_get_fb_owned_product_catalogs(
  fb_business_id = NULL,
  api_version = "v17.0",
  fields = c("id", "business", "commerce_merchant_settings", "da_display_settings",
    "default_image_url", "fallback_image_url", "feed_count", "is_catalog_segment",
    "name", "product_count", "vertical"),
  token = NULL
)
}
\arguments{
\item{fb_business_id}{Facebook business id. For details on how to find it,
see \url{https://www.facebook.com/business/help/1181250022022158}}

\item{api_version}{Defaults to "v17.0".}

\item{fields}{Defaults to all available.}

\item{token}{Facebook user token (not a page token).}
}
\description{
See the relevant page in the documentation for available fields and more details
https://developers.facebook.com/docs/marketing-api/reference/product-catalog
}
\examples{
\dontrun{
cc_get_fb_owned_product_catalogs()
}
}
