% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_insights.R
\name{cc_get_fb_page_insights}
\alias{cc_get_fb_page_insights}
\title{Get Facebook page insights}
\usage{
cc_get_fb_page_insights(
  metric = c("page_impressions"),
  start_date = NULL,
  end_date = NULL,
  api_version = "v24.0",
  cache = TRUE,
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\arguments{
\item{fb_page_token}{}
}
\description{
Official documentation: \url{https://developers.facebook.com/docs/graph-api/reference/v24.0/insights}
For avaialable metrics, see: \url{https://developers.facebook.com/docs/graph-api/reference/v24.0/insights#availmetrics}
}
