% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_page_post_insights.R
\name{cc_get_fb_page_post_insights}
\alias{cc_get_fb_page_post_insights}
\title{Get Facebook page posts insights}
\usage{
cc_get_fb_page_post_insights(
  fb_post_id,
  metrics = cc_valid_fields_fb_post_insights,
  cache = TRUE,
  update = TRUE,
  api_version = "v18.0",
  fb_page_id = NULL,
  fb_page_token = NULL
)
}
\arguments{
\item{fb_post_id}{}

\item{metrics}{}

\item{cache}{}

\item{update}{}

\item{api_version}{}

\item{fb_page_id}{}

\item{fb_page_token}{}
}
\description{
For reference, see: https://developers.facebook.com/docs/graph-api/reference/insights/#page-posts
defaults to Lifetime period for each post.
}
