% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_fb_user.R
\name{cc_get_fb_user}
\alias{cc_get_fb_user}
\title{Get Facebook user id and name, as well as other options fields}
\format{
Defaults to "data.frame". If "list", a list is returned instead;
  useful e.g. when the "picture" field is requested.
}
\usage{
cc_get_fb_user(
  fb_user_token = NULL,
  fields = c("id", "name"),
  format = "data.frame"
)
}
\arguments{
\item{fb_user_token}{Facebook user token different from page token. Can be
retrieved with [cc_get_fb_user_token()] or [cc_get_fb_long_user_token()].}
}
\value{
By default, a data frame with one row and two character columns,
  "name" and "id". Customisable with the `format` argument.
}
\description{
For details, see:
\url{https://developers.facebook.com/docs/graph-api/reference/user/}
}
\examples{
\dontrun{
cc_get_fb_user()
}
}
