% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_media.R
\name{cc_get_instagram_media}
\alias{cc_get_instagram_media}
\title{Get details about a given Instagram post based on its ig_media_id}
\usage{
cc_get_instagram_media(
  ig_media_id = NULL,
  fields = cc_valid_fields_instagram_media_v,
  api_version = "v21.0",
  ig_user_id = NULL,
  update = TRUE,
  cache = TRUE,
  token = NULL
)
}
\arguments{
\item{ig_media_id}{Instagram media identifier. A list of identifiers for your
account can be retrieved with `cc_get_instagram_media_id()`. If left to
NULL, a full list is automatically retrieved.}

\item{api_version}{Defaults to "v21.0".}

\item{ig_user_id}{Instagram user id, typically composed of 17 digits. Not to
be confused with legacy Instragram account id.}
}
\description{
It retrieves the requested fields from the APIs and introduces a few
adjustments:
}
\details{
- it always include the media id, in a column named `ig_media_id`
- if the field `owner` is included, it outputs the result as a character vector (if more than one owner, separated by ;)
- it adds a `timestamp_retrieved` column, with ISO 8601-formatted creation date in UTC
- it ensures that the output always included all requested fields, if they are valid; e.g. `is_shared_to_feed` and `media_url` may be omitted by the API (see documentation) but this function always includes the relevant column (and returns a NA value if no value is given)
- all valid fields for the given API endpoint are always requested and cached locally; only requested fields are effectively returned (but `ig_media_id` and `timestamp_retrieved` are always included as first and last column)

For details, see:
https://developers.facebook.com/docs/instagram-api/reference/ig-media
}
\examples{
\dontrun{
cc_get_instagram_media()
}
}
