% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_media_id.R
\name{cc_get_instagram_media_id}
\alias{cc_get_instagram_media_id}
\title{Get identifiers of all media published on a profile}
\usage{
cc_get_instagram_media_id(
  ig_user_id = NULL,
  api_version = "v18.0",
  token = NULL,
  max_pages = NULL,
  cache = TRUE
)
}
\arguments{
\item{ig_user_id}{Instagram user id, typically composed of 17 digits. Not to be confused with legacy Instragram account id.}

\item{api_version}{Defaults to "v18.0".}

\item{max_pages}{Defaults to NULL. If not given, it retrieves the number of media and calculates max number of pages needed. Hard-coded max number of pages at 1000.}
}
\value{
A tibble with one column named `ig_media_id` with identifiers of Instagram media.
}
\description{
Get identifiers of all media published on a profile
}
