% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_instagram_user.R
\name{cc_get_instagram_user}
\alias{cc_get_instagram_user}
\title{Get information about your Instagram user}
\usage{
cc_get_instagram_user(
  ig_user_id = NULL,
  api_version = "v19.0",
  fields = c("id", "ig_id", "username", "name", "biography", "website",
    "followers_count", "follows_count", "media_count", "profile_picture_url",
    "shopping_product_tag_eligibility"),
  fb_user_token = NULL
)
}
\arguments{
\item{ig_user_id}{Instagram user id, typically composed of 17 digits. Not to be confused with legacy Instragram account id.}

\item{api_version}{Defaults to "v18.0".}

\item{fields}{Defaults to all available, consider reducing if you don't have all relevant permissions.}

\item{fb_user_token}{Facebook user token (not a page token).}
}
\description{
If you need to retrieve your Instagram user id, consider `cc_get_instagram_user_id()`.
}
\details{
See the relevant page in the documentation for available fields and more details
https://developers.facebook.com/docs/instagram-api/reference/ig-user

Look in particular at the permissions requirements. If you have issues, consider dropping `shopping_product_tag_eligibility` from the fields, as it requires additional permissions.
}
\examples{
\dontrun{
cc_get_instagram_user()
}
}
