% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set.R
\name{cc_get_settings}
\alias{cc_get_settings}
\title{Retrieve settings for the session.}
\usage{
cc_get_settings(
  meta_api_version = NULL,
  start_date = NULL,
  end_date = NULL,
  fb_user_token = NULL,
  fb_page_token = NULL,
  fb_page_id = NULL,
  fb_business_id = NULL,
  fb_ad_account_id = NULL,
  fb_product_catalog_id = NULL,
  fb_user_id = NULL,
  ig_user_id = NULL,
  ga_email = NULL,
  ga_property_id = NULL,
  woocommerce_api_version = NULL,
  woocommerce_username = NULL,
  woocommerce_password = NULL
)
}
\arguments{
\item{start_date}{Nominally defaults to NULL, but effectively to 91 days before today.}

\item{end_date}{Nominally defaults to NULL, but effectively to yesterday.}
}
\value{
A list of named elements.
}
\description{
Typically set with `cc_set()`.
}
\examples{
\donttest{
if (interactive()) {
  dates_l <- cc_get_settings()
  dates_l
}
}

}
