% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_set.R
\name{cc_set}
\alias{cc_set}
\title{Set settings and token for the session}
\usage{
cc_set(
  meta_api_version = "v24.0",
  start_date = NULL,
  end_date = NULL,
  fb_user_token = NULL,
  fb_page_token = NULL,
  fb_page_id = NULL,
  fb_business_id = NULL,
  fb_ad_account_id = NULL,
  fb_product_catalog_id = NULL,
  fb_user_id = NULL,
  ig_user_id = NULL,
  ga_email = NULL,
  ga_property_id = NULL,
  woocommerce_api_version = "v3",
  woocommerce_username = NULL,
  woocommerce_password = NULL
)
}
\arguments{
\item{meta_api_version}{Defaults to the latest API at the time this package was
last updated. Currently, this corresponds to api version 24.0.}

\item{start_date}{Defaults to 91 days before today}

\item{end_date}{Defaults to yesterday.}

\item{fb_user_token}{Facebook user token different from page token. Can be
retrieved with [cc_get_fb_user_token()] or [cc_get_fb_long_user_token()].}

\item{fb_page_token}{Facebook page token, different from the user token. Can
be retrieved with [cc_get_fb_page_token()] or
[cc_get_fb_long_page_token()].}

\item{fb_page_id}{Facebook page identifier. Can be retrieved with
[cc_get_fb_managed_pages()].}

\item{fb_business_id}{Facebook business id. For details on how to find it,
see \url{https://www.facebook.com/business/help/1181250022022158}}

\item{ga_email}{E-mail addressed associated with relevant Google Analytics
account, passed to `googleAnalyticsR::ga_auth()`}

\item{ga_property_id}{Google Analytics property identifier. Find it with
`googleAnalyticsR::ga_account_list("ga4")`.}
}
\value{
A list of named elements.
}
\description{
Set settings and token for the session
}
\examples{
\donttest{
if (interactive()) {
  dates_l <- cc_set()
  dates_l
}
}

}
